# Conflicts:
#	src/omni_mercury_engine/federated/__init__.py
#	src/omni_mercury_engine/federated/federated_detector.py
#	src/omni_mercury_engine/infrastructure/streaming.py
#	src/omni_mercury_engine/integrations/http/client.py
#	src/omni_mercury_engine/resilience/circuit_breaker.py
#	src/omni_mercury_engine/utils/__init__.py
#	src/omni_mercury_engine/utils/resilience.py
#	src/omni_mercury_engine/utils/rng.py

+++ src/omni_mercury_engine/api/routes/detection.py
@@ -405,7 +405,9 @@ async def detect_three_r(
             left_score = compute_recursion_score(x[:mid], depth - 1)
             right_score = compute_recursion_score(x[mid:], depth - 1)
 
-            return 0.5 * (left_score + right_score) + 0.5 * np.std(x) / (np.mean(np.abs(x)) + 1e-8)
+            return float(
+                0.5 * (left_score + right_score) + 0.5 * np.std(x) / (np.mean(np.abs(x)) + 1e-8)
+            )
 
         recursion_score = compute_recursion_score(data, request.recursion_depth)
         recursion_score = float(np.clip(recursion_score, 0.0, 1.0))
